# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "T-Rader API",
    "description": "日本株投資プラットフォーム T-Rader の API",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/strategies": {
      "get": {
        "tags": [
          "strategies"
        ],
        "summary": "戦略一覧",
        "operationId": "list_strategies",
        "responses": {
          "200": {
            "description": "戦略一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Strategy"
                  }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "strategies"
        ],
        "summary": "戦略作成",
        "operationId": "create_strategy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStrategyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Strategy"
                }
              }
            }
          },
          "400": {
            "description": "リクエストパラメータが不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type ヘッダが application/json ではない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "リクエストボディのパースに失敗",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/strategies/{id}/risk-policy": {
      "get": {
        "tags": [
          "strategies"
        ],
        "summary": "戦略の銘柄集中度上限 (`max_position_ratio`) を取得",
        "operationId": "get_risk_policy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "戦略 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyRiskPolicyResponse"
                }
              }
            }
          },
          "400": {
            "description": "リクエストパラメータが不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "strategies"
        ],
        "summary": "戦略の銘柄集中度上限 (`max_position_ratio`) を更新。`null` で上限を解除する",
        "operationId": "put_risk_policy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "戦略 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutStrategyRiskPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyRiskPolicyResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type ヘッダが application/json ではない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateStrategyRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S"
          },
          "sort_order": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "API エラーレスポンスの JSON 構造",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "エラーメッセージ"
          }
        }
      },
...
      "PutStrategyRiskPolicyRequest": {
        "type": "object",
        "properties": {
          "max_position_ratio": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "銘柄の保有時価 / 戦略の投資可能額 の上限比率。(0, 1] の範囲。`null` で上限を解除する"
          }
        },
        "additionalProperties": false
      },
...
      "Strategy": {
        "type": "object",
        "required": [
          "id",
          "name",
          "sort_order",
          "created_at",
          "updated_at",
          "agents_md",
          "skills",
          "agent_graph",
          "risk_policy"
        ],
        "properties": {
          "agent_graph": {
            "type": "string"
          },
          "agents_md": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "risk_policy": {
            "$ref": "#/components/schemas/Value"
          },
          "skills": {
            "$ref": "#/components/schemas/Value"
          },
          "sort_order": {
            "type": "integer",
            "format": "int32"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "StrategyRiskPolicyResponse": {
        "type": "object",
        "properties": {
          "max_position_ratio": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          }
        }
      },
...
      "Value": {},
...
    }
  },
...
}